fix: split traits
